refactoring

+++ src/Principal/SensorUltrasonico.h
@@ -12,7 +12,7 @@
 class SensorUltrasonico: public Thread
 {
 public:
-        SensorUltrasonico(uint8_t pinoEcho, uint8_t pinoTrigger, uint16_t frequencia, int intervalo);
+        SensorUltrasonico(uint8_t pinoEcho, uint8_t pinoTrigger, uint16_t frequencia, uint16_t intervalo);
         bool existeObjeto(void);
         void detectar(void);
 
@@ -22,7 +22,7 @@ private:
       	uint8_t pinoTrigger;
       	unsigned long distanciaEcoada;
       	uint16_t frequencia;
-      	int intervalo;
+      	uint16_t intervalo;
       	uint8_t deteccoes;
       	bool objetoDetectado;
